Blocked out UI for adult page and added isAdult field to the dealers
There's not much to see on the adult page since it's all plain-old, PG placeholders, but the UI is there according to igglypou's sketch.

I also added an isAdult field to the live dealer data and filtered against it on the main dealers page. I have verified that if a dealer is an adult dealer, it will not show up on the main dealers page.