Caught a bug in  the function that probes the payload from the donation page
After donating through the live RegFox page to test it, I noticed my donation wasn't showing up. It turns out that my probe function wasn't finding the field for whether or not the registrant covered the CC fee. However, this function is essentially the same as the probe in the main registration hook, and the logic is fine otherwise. So, I just removed the test for the CC fee field since it is already taken care of and  working in the `getDonationAmountFromRegistrant` function.